requirements.txt
decord==0.6.0
einops==0.8.1
Flask==3.1.2
ftfy==6.3.1
huggingface_hub==1.1.5
hydra-core==1.3.2
iopath==0.1.10
matplotlib==3.10.7
numba==0.62.1
numpy==2.3.5
omegaconf==2.3.0
openai==2.8.1
opencv_python==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
psutil==7.0.0
pycocotools==2.0.10
pytest==9.0.1
PyYAML==6.0.2
regex==2025.11.3
scikit_learn==1.7.2
scipy==1.16.3
scikit-image==0.24.0
submitit==1.5.3
tensordict==0.10.0
tidecv==1.0.1
timm==1.0.22
torch==2.7.1+cu126
torchcodec==0.8.1
torchmetrics==1.8.2
torchvision==0.22.1+cu126
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.4.8